refactor(suite-native): vault detail screen
